+++ src/components/OrderButtons.tsx
@@ -6,7 +6,6 @@ import { OrderSubmissionParameters } from "../types/OrderSubmissionParameters";
 import { boundingBoxMatch, filterAddWildcards } from "../utils/CMR";
 import { IEnvironment } from "../utils/environment";
 import { hasChanged } from "../utils/hasChanged";
-import { UserContext } from "../utils/state";
 import { EarthdataFlow } from "./EarthdataFlow";
 import { ScriptButton } from "./ScriptButton";
 import { SubmitButton } from "./SubmitButton";
@@ -25,7 +24,6 @@ interface IOrderButtonsState {
 }
 
 export class OrderButtons extends React.Component<IOrderButtonsProps, IOrderButtonsState> {
-  public static contextType = UserContext;
 
   public constructor(props: IOrderButtonsProps) {
     super(props);