+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Remove "Order Files" button in preparation for the decomissioning of the data
   ordering system (hermes).
-* Change the "Large/custom order" button text -> "Earthdata Order". Update
+* Change the "Large/custom order" button text -> "Order Data". Update
   tooltip/popup description to no longer reference file order limits. This
   langauge no longer makes sense without the context of the removed "Order
   Files" button.