+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v3.0.0 (TBD)
+
+* Remove "Order Files" button in preparation for the decomissioning of the data
+  ordering system (hermes).
+
 # v2.1.0 (2024-06-24)
 
 * Add support for and prefer cloud-hosted datasets to the "Large/Custom Order"